third_party/WebKit/LayoutTests/editing/selection/extend/go-out-of-readonly-textarea.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<script>
test(() => {
  assert_selection(
    '<div><textarea readonly="readonly">|foo</textarea></div>',
    selection => selection.extend(selection.document.querySelector('div'), 0),
    '<div><textarea readonly="readonly">|foo</textarea></div>');
});
</script>
